Move BreakingNews related class in a BreakingNews package
